Replace calls File::Copy::Recursive subroutines.
This module uses at least one of three subroutines, fcopy(), dircopy() and
rcopy() exported by CPAN distribution File-Copy-Recursive (FCR).  As of
version 0.40 of that library, a bug prevents it from building, testing and
being installed successfully on FreeBSD and other operating systems.  This
means that all other CPAN libraries which depend on FCR cannot be installed
(at least not without resort to 'force install') on those systems.  This
affects approximately one-fifth of all distributions on CPAN.

CPAN library File-Copy-Recursive-Reduced (FCR2) has been developed for the
purpose of providing drop-in replacements for a significant portion of all
usages of those three FCR functions.  This commit replaces FCR functions with
their equivalent FCR2 functions.  This library's prerequisites are now
satisfied and it can be built, tested and installed on FreeBSD and other
affected platforms.  That in turn means that other CPAN distributions
dependent on this one can also be tested and install on those platforms.